# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for transport booking",
        "title": "SimbirGO REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alexander Soldatov",
            "email": "soldatovalex207z@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/api/Admin/Rent": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание аренды транспорта с id = transportId пользователем с id = userId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "Создание новой аренды",
                "parameters": [
                    {
                        "description": "Rent data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rentHandler.AdminCreateRent.rentData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/Admin/Rent/End/{rentId}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Завершение аренды с id = {rentId}.\nПроисходит рассчет итоговой суммы аренды и если она оказывается больше, чем сумма на счете пользователя, то в завершить аренду нельзя.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "Завершение аренды",
                "parameters": [
                    {
                        "type": "number",
                        "description": "lat",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "long",
                        "name": "long",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/Admin/Rent/{rentId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение информации об аренде с id = {rentId}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "Информации об аренде",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновление информации об аренде с id = {rentId}\nЕсли в обновлении аренды указывается дата ее окончания, то аренда считается завершенной.\nПроисходит рассчет итоговой суммы аренды и если она оказывается больше, чем сумма на счете пользователя, то обновить аренду нельзя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "Обновление аренды",
                "parameters": [
                    {
                        "description": "Rent data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rentHandler.AdminUpdateRent.rentData"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Rent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаление аренды с id = {rentId}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "Удаление аренды",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/Admin/TransportHistory/{transportId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение информации обо всех арендах транспорта с id = {transportId}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "История аренды транспорта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transport id",
                        "name": "transportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Rent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/Admin/UserHistory/{userId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение информации обо всех арендах пользователем с id = {userId}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminRentController"
                ],
                "summary": "История аренды пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Rent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.Rent": {
            "type": "object",
            "properties": {
                "finalPrice": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "priceOfUnit": {
                    "type": "number"
                },
                "priceType": {
                    "type": "string",
                    "enum": [
                        "Minutes",
                        " Days"
                    ]
                },
                "timeEnd": {
                    "type": "string"
                },
                "timeStart": {
                    "type": "string"
                },
                "transportId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "httpUtil.ResponseError": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string",
                    "example": "error occures"
                }
            }
        },
        "rentHandler.AdminCreateRent.rentData": {
            "type": "object",
            "required": [
                "priceOfUnit",
                "priceType",
                "timeStart",
                "transportId",
                "userId"
            ],
            "properties": {
                "priceOfUnit": {
                    "type": "number"
                },
                "priceType": {
                    "type": "string"
                },
                "timeEnd": {
                    "type": "string"
                },
                "timeStart": {
                    "type": "string"
                },
                "transportId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}